FAQ file references
The FAQ has a lot of good info and is well written. Any reader of this quick readme should know about the documentation within the FAQ file. I could not figure out a way to universally link to it within githubs flavor of .md markdown, so I just referenced it by textual explanation.