No longer read unnecessary data in -readFanData
-readFanData was always reading the temperature and the speeds of all
of the fans regardless of what data was displayed in the menu bar. Now,
-readFanData only reads the data necessary to update the text in the
menubar, so, if only the temp is displayed, only the temp is read. The
updating of the fan speeds in the menu is now only done when the user
clicks on the menu. Also: cleaned up some test code and deleted an
unnecessary comment.